Separated out v9 and v10 protocols into their own files
